# Conflicts:
#	include/darmok/protobuf.hpp
#	src/program_core.cpp
#	src/scene_assimp.cpp
#	src/texture.cpp

+++ src/asset_core.cpp
@@ -1246,7 +1246,7 @@ namespace darmok
 
         try
         {
-            cli.parse(args.size(), args.data());
+            cli.parse(static_cast<int>(args.size()), args.data());
             cfg.fix(cli);
 
             if (cfg.dry)
